Improvements to handling of text attribute variables
We now ensure they are *always* set, either to actual values based on
the output of tput, or the empty string if we ran with --no-colour. We
should have always done this, as otherwise we'll hit unbound variable
errors if we reference them in "no colour" mode. This necessitates that
we run colour_init() on script startup, though we handle the case where
the script writer forgets to do so (at least, within our own logic).